ISO: Reorganize tests
